+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Renderer Class
 - Add logger support
 - Added CLI interface, reads scene from cli args
-- Sphere Intersection returns Ray 
+- Sphere Intersection returns Ray
+- added support for simple lights 
 
 ## [0.1.0]
 ### Added